app/pyproject.toml
[project]
name = "api-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "jinja2>=3.1.6",
    "openapi-spec-validator>=0.7.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[build-system]
requires = ["setuptools>=67.0.0", "wheel>=0.41.0"]
build-backend = "setuptools.build_meta"
